requirements.txt
PyQt6
Crypto
scipy
folium
geopy
whirlpool
matplotlib
numpy
asn1crypto
base45
base58
gostcrypto
blake3
gmssl
twofish
pgpy
pandas
pdf2image
pyminizip
bcrypt
pyzipper
capstone
pefile
uncompyle6
msgpack
pillow
cryptography
pycryptodome
oqs
kyber_py
jwt
qrcode
cffi
# pip install --no-build-isolation python-camellia (install separately, bc of deprecated version)
des
androguard
pyexiftool
arc4